+++ config/validate.go
@@ -13,7 +13,7 @@ import (
 	"github.com/rs/zerolog/log"
 )
 
-func (c Config) validate() []string {
+func (c *Config) Validate() []string {
 	var validationErrors []string
 	log.Debug().Msg("Validating config file...")
 
@@ -25,6 +25,13 @@ func (c Config) validate() []string {
 	// Validate App Mode
 	if results := c.validateAppMode(); len(results) > 0 {
 		validationErrors = append(validationErrors, results...)
+	}
+
+	// Validate API settings
+	if c.App.API.Enabled {
+		if results := c.validateAPI(); len(results) > 0 {
+			validationErrors = append(validationErrors, results...)
+		}
 	}
 
 	// Validate Frigate polling method
@@ -130,14 +137,29 @@ func (c *Config) validateAppMode() []string {
 	if strings.ToLower(c.App.Mode) != "events" && strings.ToLower(c.App.Mode) != "reviews" {
 		appErrors = append(appErrors, "MQTT mode must be 'events' or 'reviews'")
 	}
-	if c.Frigate.Version < 14 && strings.ToLower(c.App.Mode) == "reviews" {
+	if Internal.FrigateVersion < 14 && strings.ToLower(c.App.Mode) == "reviews" {
 		appErrors = append(appErrors, "Frigate must be version 0.14 or higher to use 'reviews' mode. Please use 'events' mode or update Frigate.")
 	}
 	log.Debug().Msgf("App mode: %v", c.App.Mode)
 	return appErrors
 }
 
-func (c Config) validateFrigatePolling() []string {
+func (c *Config) validateAPI() []string {
+	var apiErrors []string
+
+	// Set default port if needed
+	if c.App.API.Port == 0 {
+		c.App.API.Port = 8000
+	}
+
+	if c.App.API.Port <= 0 || c.App.API.Port > 65535 {
+		apiErrors = append(apiErrors, "Invalid API port")
+	}
+
+	return apiErrors
+}
+
+func (c *Config) validateFrigatePolling() []string {
 	var pollingErrors []string
 	webapi := c.Frigate.WebAPI.Enabled
 	mqtt := c.Frigate.MQTT.Enabled
@@ -169,7 +191,7 @@ func (c Config) validateFrigatePolling() []string {
 	return pollingErrors
 }
 
-func (c Config) validateFrigateConnectivity() []string {
+func (c *Config) validateFrigateConnectivity() []string {
 	var response []byte
 	var err error
 	var connectivityErrors []string
@@ -213,6 +235,7 @@ func (c Config) validateFrigateConnectivity() []string {
 	for current_attempt < max_attempts {
 		response, err = util.HTTPGet(statsAPI, c.Frigate.Insecure, "", c.Frigate.Headers...)
 		if err != nil {
+			Internal.Status.Frigate.API = "unreachable"
 			log.Warn().
 				Err(err).
 				Int("attempt", current_attempt).
@@ -226,22 +249,25 @@ func (c Config) validateFrigateConnectivity() []string {
 		}
 	}
 	if current_attempt == max_attempts {
-		log.Fatal().
+		Internal.Status.Frigate.API = "unreachable"
+		log.Error().
 			Err(err).
 			Msgf("Max attempts reached - Cannot reach Frigate server at %v", url)
+		connectivityErrors = append(connectivityErrors, "Max attempts reached - Cannot reach Frigate server at "+url)
 	}
 	var stats models.FrigateStats
 	json.Unmarshal([]byte(response), &stats)
 	log.Info().Msgf("Successfully connected to %v", url)
+	Internal.Status.Frigate.API = "ok"
 	if stats.Service.Version != "" {
 		log.Debug().Msgf("Frigate server is running version %v", stats.Service.Version)
 		// Save major version number
-		c.Frigate.Version, _ = strconv.Atoi(strings.Split(stats.Service.Version, ".")[1])
+		Internal.FrigateVersion, _ = strconv.Atoi(strings.Split(stats.Service.Version, ".")[1])
 	}
 	return connectivityErrors
 }
 
-func (c Config) validateMQTT() []string {
+func (c *Config) validateMQTT() []string {
 	var configErrors []string
 	// Check MQTT Config
 	log.Debug().Msg("MQTT Enabled.")
@@ -258,7 +284,7 @@ func (c Config) validateMQTT() []string {
 	return configErrors
 }
 
-func (c Config) validateCameraExclusions() {
+func (c *Config) validateCameraExclusions() {
 	// Check for camera exclusions
 	if len(c.Frigate.Cameras.Exclude) > 0 {
 		log.Debug().Msg("Cameras to exclude from alerting:")
@@ -268,7 +294,7 @@ func (c Config) validateCameraExclusions() {
 	}
 }
 
-func (c Config) validateQuietHours() []string {
+func (c *Config) validateQuietHours() []string {
 	var quietHoursErrors []string
 	// Check quiet hours config
 	if c.Alerts.Quiet.Start != "" || c.Alerts.Quiet.End != "" {
@@ -290,7 +316,7 @@ func (c Config) validateQuietHours() []string {
 	return quietHoursErrors
 }
 
-func (c Config) validateAlertGeneral() []string {
+func (c *Config) validateAlertGeneral() []string {
 	var alertErrors []string
 	// Check action on no snapshot available
 	if strings.ToLower(c.Alerts.General.NoSnap) != "allow" && strings.ToLower(c.Alerts.General.NoSnap) != "drop" {
@@ -313,7 +339,7 @@ func (c Config) validateAlertGeneral() []string {
 	return alertErrors
 }
 
-func (c Config) validateZoneFilters() []string {
+func (c *Config) validateZoneFilters() []string {
 	var filterErrors []string
 	// Check Zone filtering config
 	if strings.ToLower(c.Alerts.Zones.Unzoned) != "allow" && strings.ToLower(c.Alerts.Zones.Unzoned) != "drop" {
@@ -341,7 +367,7 @@ func (c Config) validateZoneFilters() []string {
 	return filterErrors
 }
 
-func (c Config) validateLabelFiltering() []string {
+func (c *Config) validateLabelFiltering() []string {
 	var labelErrors []string
 	// Check Label filtering config
 	if c.Alerts.Labels.MinScore > 0 {
@@ -384,9 +410,11 @@ func (c Config) validateLabelFiltering() []string {
 	return labelErrors
 }
 
-func (c Config) validateDiscord() []string {
+func (c *Config) validateDiscord() []string {
 	var discordErrors []string
 	log.Debug().Msg("Discord alerting enabled.")
+	discordStatus := models.NotifierStatus{Enabled: true, Status: "configured, not used yet"}
+	Internal.Status.Notifications.Discord = append(Internal.Status.Notifications.Discord, discordStatus)
 	if c.Alerts.Discord.Webhook == "" {
 		discordErrors = append(discordErrors, "No Discord webhook specified!")
 	}
@@ -397,9 +425,11 @@ func (c Config) validateDiscord() []string {
 	return discordErrors
 }
 
-func (c Config) validateGotify() []string {
+func (c *Config) validateGotify() []string {
 	var gotifyErrors []string
 	log.Debug().Msg("Gotify alerting enabled.")
+	gotifyStatus := models.NotifierStatus{Enabled: true, Status: "configured, not used yet"}
+	Internal.Status.Notifications.Gotify = append(Internal.Status.Notifications.Gotify, gotifyStatus)
 	if c.Alerts.Gotify.Server == "" {
 		gotifyErrors = append(gotifyErrors, "No Gotify server specified!")
 	}
@@ -418,9 +448,11 @@ func (c Config) validateGotify() []string {
 	return gotifyErrors
 }
 
-func (c Config) validateSMTP() []string {
+func (c *Config) validateSMTP() []string {
 	var smtpErrors []string
 	log.Debug().Msg("SMTP alerting enabled.")
+	smtpStatus := models.NotifierStatus{Enabled: true, Status: "configured, not used yet"}
+	Internal.Status.Notifications.SMTP = append(Internal.Status.Notifications.SMTP, smtpStatus)
 	if c.Alerts.SMTP.Server == "" {
 		smtpErrors = append(smtpErrors, "No SMTP server specified!")
 	}
@@ -445,9 +477,11 @@ func (c Config) validateSMTP() []string {
 	return smtpErrors
 }
 
-func (c Config) validateTelegram() []string {
+func (c *Config) validateTelegram() []string {
 	var telegramErrors []string
 	log.Debug().Msg("Telegram alerting enabled.")
+	telegramStatus := models.NotifierStatus{Enabled: true, Status: "configured, not used yet"}
+	Internal.Status.Notifications.Telegram = append(Internal.Status.Notifications.Telegram, telegramStatus)
 	if c.Alerts.Telegram.ChatID == 0 {
 		telegramErrors = append(telegramErrors, "No Telegram Chat ID specified!")
 	}
@@ -461,9 +495,11 @@ func (c Config) validateTelegram() []string {
 	return telegramErrors
 }
 
-func (c Config) validatePushover() []string {
+func (c *Config) validatePushover() []string {
 	var pushoverErrors []string
 	log.Debug().Msg("Pushover alerting enabled.")
+	pushoverStatus := models.NotifierStatus{Enabled: true, Status: "configured, not used yet"}
+	Internal.Status.Notifications.Pushover = append(Internal.Status.Notifications.Pushover, pushoverStatus)
 	if c.Alerts.Pushover.Token == "" {
 		pushoverErrors = append(pushoverErrors, "No Pushover API token specified!")
 	}
@@ -493,9 +529,11 @@ func (c Config) validatePushover() []string {
 	return pushoverErrors
 }
 
-func (c Config) validateNtfy() []string {
+func (c *Config) validateNtfy() []string {
 	var ntfyErrors []string
 	log.Debug().Msg("Ntfy alerting enabled.")
+	ntfyStatus := models.NotifierStatus{Enabled: true, Status: "configured, not used yet"}
+	Internal.Status.Notifications.Ntfy = append(Internal.Status.Notifications.Ntfy, ntfyStatus)
 	if c.Alerts.Ntfy.Server == "" {
 		ntfyErrors = append(ntfyErrors, "No Ntfy server specified!")
 	}
@@ -515,9 +553,11 @@ func (c Config) validateNtfy() []string {
 	return ntfyErrors
 }
 
-func (c Config) validateWebhook() []string {
+func (c *Config) validateWebhook() []string {
 	var webhookErrors []string
 	log.Debug().Msg("Webhook alerting enabled.")
+	webhookStatus := models.NotifierStatus{Enabled: true, Status: "configured, not used yet"}
+	Internal.Status.Notifications.Webhook = append(Internal.Status.Notifications.Webhook, webhookStatus)
 	if c.Alerts.Webhook.Server == "" {
 		webhookErrors = append(webhookErrors, "No Webhook server specified!")
 	}
@@ -529,7 +569,7 @@ func (c Config) validateWebhook() []string {
 	return webhookErrors
 }
 
-func (c Config) validateAlertingEnabled() string {
+func (c *Config) validateAlertingEnabled() string {
 	// Check to ensure at least one alert provider is configured
 	if c.Alerts.Discord.Enabled {
 		return ""
@@ -555,7 +595,7 @@ func (c Config) validateAlertingEnabled() string {
 	return "No alerting methods have been configured. Please check config file syntax!"
 }
 
-func (c Config) validateAppMonitoring() []string {
+func (c *Config) validateAppMonitoring() []string {
 	var monitoringErrors []string
 	log.Debug().Msg("App monitoring enabled.")
 	if c.Monitor.URL == "" {