+++ config/validate_test.go
@@ -2,14 +2,16 @@ package config
 
 import (
 	"testing"
+
+	"github.com/0x2142/frigate-notify/models"
 )
 
 func TestValidateAppMode(t *testing.T) {
-	config := Config{Frigate: &Frigate{}}
+	config := Config{Frigate: &models.Frigate{}}
 
 	// Check good config
 	config.App.Mode = "reviews"
-	config.Frigate.Version = 14
+	Internal.FrigateVersion = 14
 	result := config.validateAppMode()
 	expected := 0
 	if len(result) != expected {
@@ -25,7 +27,7 @@ func TestValidateAppMode(t *testing.T) {
 
 	// Check incompatible version
 	config.App.Mode = "reviews"
-	config.Frigate.Version = 13
+	Internal.FrigateVersion = 13
 	result = config.validateAppMode()
 	expected = 1
 	if len(result) != expected {
@@ -33,8 +35,38 @@ func TestValidateAppMode(t *testing.T) {
 	}
 }
 
+func TestValidateAPI(t *testing.T) {
+	config := Config{App: models.App{}}
+
+	config.App.API.Enabled = true
+
+	// Validate default port set
+	config.validateAPI()
+	if config.App.API.Port != 8000 {
+		t.Errorf("Expected: 80, Got: %v", config.App.API.Port)
+
+	}
+
+	// Check good config
+	config.App.API.Port = 8080
+	result := config.validateAPI()
+	expected := 0
+	if len(result) != expected {
+		t.Errorf("Expected: %v error(s), Got: %v", expected, result)
+	}
+
+	// Check bad config
+	config.App.API.Port = 65540
+	result = config.validateAPI()
+	expected = 1
+	if len(result) != expected {
+		t.Errorf("Expected: %v error(s), Got: %v", expected, result)
+	}
+
+}
+
 func TestValidateFrigatePolling(t *testing.T) {
-	config := Config{Frigate: &Frigate{}}
+	config := Config{Frigate: &models.Frigate{}}
 
 	// Test one method configured
 	config.Frigate.MQTT.Enabled = true
@@ -62,7 +94,7 @@ func TestValidateFrigatePolling(t *testing.T) {
 }
 
 func TestValidateMQTT(t *testing.T) {
-	config := Config{Frigate: &Frigate{}}
+	config := Config{Frigate: &models.Frigate{}}
 
 	// Test correct config
 	config.Frigate.MQTT.Enabled = true
@@ -94,7 +126,7 @@ func TestValidateMQTT(t *testing.T) {
 }
 
 func TestValidateQuietHours(t *testing.T) {
-	config := Config{Alerts: &Alerts{}}
+	config := Config{Alerts: &models.Alerts{}}
 
 	// Test valid config
 	config.Alerts.Quiet.Start = "03:50"
@@ -123,7 +155,7 @@ func TestValidateQuietHours(t *testing.T) {
 }
 
 func TestValidateAlertGeneral(t *testing.T) {
-	config := Config{Alerts: &Alerts{}}
+	config := Config{Alerts: &models.Alerts{}}
 
 	// Test valid config
 	config.Alerts.General.NoSnap = "allow"
@@ -143,7 +175,7 @@ func TestValidateAlertGeneral(t *testing.T) {
 }
 
 func TestValidateDiscord(t *testing.T) {
-	config := Config{Alerts: &Alerts{}}
+	config := Config{Alerts: &models.Alerts{}}
 
 	// Test valid config
 	config.Alerts.Discord.Webhook = "https://something.test"
@@ -163,7 +195,7 @@ func TestValidateDiscord(t *testing.T) {
 }
 
 func TestValidateGotify(t *testing.T) {
-	config := Config{Alerts: &Alerts{}}
+	config := Config{Alerts: &models.Alerts{}}
 
 	// Test valid config
 	config.Alerts.Gotify.Server = "https://something.test"
@@ -192,7 +224,7 @@ func TestValidateGotify(t *testing.T) {
 }
 
 func TestValidateSMTP(t *testing.T) {
-	config := Config{Alerts: &Alerts{}}
+	config := Config{Alerts: &models.Alerts{}}
 
 	// Test valid config
 	config.Alerts.SMTP.Server = "192.0.2.10"
@@ -241,7 +273,7 @@ func TestValidateSMTP(t *testing.T) {
 }
 
 func TestValidateTelegram(t *testing.T) {
-	config := Config{Alerts: &Alerts{}}
+	config := Config{Alerts: &models.Alerts{}}
 
 	// Test valid config
 	config.Alerts.Telegram.ChatID = 1234
@@ -270,7 +302,7 @@ func TestValidateTelegram(t *testing.T) {
 }
 
 func TestValidatePushover(t *testing.T) {
-	config := Config{Alerts: &Alerts{}}
+	config := Config{Alerts: &models.Alerts{}}
 
 	// Test valid config
 	config.Alerts.Pushover.Token = "abcd"
@@ -325,7 +357,7 @@ func TestValidatePushover(t *testing.T) {
 }
 
 func TestValidateNtfy(t *testing.T) {
-	config := Config{Alerts: &Alerts{}}
+	config := Config{Alerts: &models.Alerts{}}
 
 	// Test valid config
 	config.Alerts.Ntfy.Server = "https://ntfy.test"
@@ -353,7 +385,7 @@ func TestValidateNtfy(t *testing.T) {
 	}
 }
 func TestValidateWebhook(t *testing.T) {
-	config := Config{Alerts: &Alerts{}}
+	config := Config{Alerts: &models.Alerts{}}
 
 	// Test valid config
 	config.Alerts.Webhook.Server = "https://webhook.test"
@@ -374,7 +406,7 @@ func TestValidateWebhook(t *testing.T) {
 }
 
 func TestValidateAlertingEnabled(t *testing.T) {
-	config := Config{Alerts: &Alerts{}}
+	config := Config{Alerts: &models.Alerts{}}
 
 	// Test valid config
 	config.Alerts.Discord.Enabled = true