+++ events/mqtt.go
@@ -14,9 +14,12 @@ import (
 )
 
 var mqtt_topic string
+var client mqtt.Client
 
 // SubscribeMQTT establishes subscription to MQTT server & listens for messages
 func SubscribeMQTT() {
+	config.Internal.Status.Health = "frigate mqtt connecting"
+	config.Internal.Status.Frigate.MQTT = "connecting"
 	mqtt_topic = fmt.Sprintf("%s/%s", config.ConfigData.Frigate.MQTT.TopicPrefix, strings.ToLower(config.ConfigData.App.Mode))
 	// MQTT client configuration
 	mqttServer := fmt.Sprintf("tcp://%s:%d", config.ConfigData.Frigate.MQTT.Server, config.ConfigData.Frigate.MQTT.Port)
@@ -43,14 +46,19 @@ func SubscribeMQTT() {
 	var subscribed = false
 	var retry = 0
 	for !subscribed {
+		config.Internal.Status.Health = "frigate mqtt unable to connect"
+		config.Internal.Status.Frigate.MQTT = "unreachable"
 		if retry >= 3 {
 			log.Fatal().Msgf("Max retries exceeded. Failed to establish MQTT session to %s", config.ConfigData.Frigate.MQTT.Server)
 		}
 		// Connect to MQTT broker
-		client := mqtt.NewClient(opts)
+		client = mqtt.NewClient(opts)
 
 		if token := client.Connect(); token.Wait() && token.Error() != nil {
 			retry += 1
+			config.Internal.Status.Health = "frigate mqtt unable to connect"
+			config.Internal.Status.Frigate.MQTT = "unreachable"
+
 			log.Warn().Msgf("Could not connect to MQTT at %v: %v", config.ConfigData.Frigate.MQTT.Server, token.Error())
 			log.Warn().Msgf("Retrying in 10 seconds. Attempt %v of 3.", retry)
 			time.Sleep(10 * time.Second)
@@ -60,8 +68,17 @@ func SubscribeMQTT() {
 	}
 }
 
+// disconnectMQTT simply disconnects the MQTT client
+func DisconnectMQTT() {
+	log.Info().Msg("Ending MQTT session")
+	client.Disconnect(300)
+	log.Info().Msg("MQTT disconnected")
+}
+
 // connectionLostHandler logs error message on MQTT connection loss
 func connectionLostHandler(c mqtt.Client, err error) {
+	config.Internal.Status.Health = "frigate mqtt connection lost"
+	config.Internal.Status.Frigate.MQTT = "unreachable"
 	log.Error().
 		Err(err).
 		Msg("Lost connection to MQTT broker")
@@ -70,7 +87,11 @@ func connectionLostHandler(c mqtt.Client, err error) {
 // connectHandler logs message on MQTT connection
 func connectHandler(client mqtt.Client) {
 	log.Info().Msg("Connected to MQTT.")
+	config.Internal.Status.Health = "ok"
+	config.Internal.Status.Frigate.MQTT = "ok"
 	if subscription := client.Subscribe(mqtt_topic, 0, handleMQTTMsg); subscription.Wait() && subscription.Error() != nil {
+		config.Internal.Status.Health = "frigate mqtt unable to subscribe"
+		config.Internal.Status.Frigate.MQTT = "unreachable"
 		log.Error().Msgf("Failed to subscribe to topic: %s", mqtt_topic)
 		time.Sleep(10 * time.Second)
 	}